SAX API Refactoring
- moved inner classes into classes of their own
- some refactoring to remove duplicate code

DAFFODIL-2422
